+++ tests/Feature/ApiDataImporterTest.php
@@ -464,7 +464,7 @@ class ApiDataImporterTest extends TestCase {
         ])
             ->assertStatus(200)
             ->assertJson([
-                'errors' => ['[2] Attribute could not be imported: Abmessungen'],
+                'errors' => ['[2] Attribute could not be imported: {{Abmessungen}} => {{1,2,3,cm}}'],
                 'summary' => [
                     'create' => 1,
                     'update' => 0,
@@ -472,7 +472,29 @@ class ApiDataImporterTest extends TestCase {
                 ]
             ]);
     }
+    
+    public function testValidationWithDuplicateColumnMapping(){
+        $response = $this->userRequest()->post('/api/v1/entity/import/validate', [
+            'file' => $this->createDefaultCSVFile(),
+            'data' => $this->getData(
+                attributes: [
+                    13 => "Notizen", // Notizen
+                    19 => "Notizen", // Aufbewahrung
+                ]
+            ),
+            'metadata' => $this->getMetaData(),
+        ]);
+        $response->assertStatus(200);
 
+        $response->assertJson([
+            'errors' => [],
+            'summary' => [
+                'create' => 4,
+                'update' => 1,
+                'conflict' => 0
+            ]
+        ]);
+    }
 
     public function testDataImportSuccess() {
         $response = $this->userRequest()->post('/api/v1/entity/import', [
@@ -576,5 +598,21 @@ class ApiDataImporterTest extends TestCase {
                 'on_name' => null
             ]
         ]);
+    }
+    
+    public function testDataImportFailsWithDuplicateColumnMapping(){
+        $response = $this->userRequest()->post('/api/v1/entity/import', [
+            'file' => $this->createDefaultCSVFile(),
+            'data' => $this->getData(
+                attributes: [
+                    13 => "Notizen", // Notizen
+                    19 => "Notizen", // Aufbewahrung
+                ]
+            ),
+            'metadata' => $this->getMetaData(),
+        ])
+            ->assertStatus(201);
+
+            
     }
 }